requirements.txt
absl-py==0.9.0
appnope==0.1.0
astor==0.8.1
attrs==19.3.0
aws-cdk.assets==1.39.0
aws-cdk.aws-cloudformation==1.39.0
aws-cdk.aws-cloudwatch==1.39.0
aws-cdk.aws-ec2==1.39.0
aws-cdk.aws-ecr==1.39.0
aws-cdk.aws-events==1.39.0
aws-cdk.aws-iam==1.39.0
aws-cdk.aws-kms==1.39.0
aws-cdk.aws-lambda==1.39.0
aws-cdk.aws-logs==1.39.0
aws-cdk.aws-s3==1.39.0
aws-cdk.aws-s3-assets==1.39.0
aws-cdk.aws-sns==1.39.0
aws-cdk.aws-sqs==1.39.0
aws-cdk.aws-ssm==1.39.0
aws-cdk.cdk-assets-schema==1.39.0
aws-cdk.cloud-assembly-schema==1.39.0
aws-cdk.core==1.39.0
aws-cdk.custom-resources==1.39.0
aws-cdk.cx-api==1.39.0
aws-cdk.region-info==1.39.0
backcall==0.1.0
bleach==3.1.4
cachetools==4.1.0
cattrs==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
constructs==3.0.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
editdistance==0.5.3
efficientnet==1.0.0
entrypoints==0.3
essential-generators==0.9.2
fonttools==4.8.1
gast==0.2.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
imageio==2.8.0
imgaug==0.4.0
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
jsii==1.5.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
keras-ocr==0.8.3
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
notebook==6.0.3
numpy==1.18.3
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
publication==0.0.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.1.0.post3
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
PyWavelets==1.1.1
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
Shapely==1.7.0
six==1.14.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
typing-extensions==3.7.4.2
urllib3==1.25.9
validators==0.14.3
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
